+++ handler/packfileUpdated.go
@@ -6,9 +6,11 @@ import (
 	"path"
 	"path/filepath"
 	"strconv"
+	"sync"
 
 	"github.com/gitopia/gitopia-go/logger"
 	"github.com/gitopia/gitopia-storage/app"
+	"github.com/gitopia/gitopia-storage/handler/storage"
 	"github.com/gitopia/gitopia-storage/utils"
 	"github.com/pkg/errors"
 	"github.com/sirupsen/logrus"
@@ -93,14 +95,14 @@ func UnmarshalPackfileUpdatedEvent(eventBuf []byte) ([]PackfileUpdatedEvent, err
 }
 
 type PackfileUpdatedEventHandler struct {
-	gc           *app.GitopiaProxy
-	pinataClient *PinataClient
+	gc             *app.GitopiaProxy
+	storageManager *storage.Manager
 }
 
-func NewPackfileUpdatedEventHandler(g *app.GitopiaProxy, pinataClient *PinataClient) PackfileUpdatedEventHandler {
+func NewPackfileUpdatedEventHandler(g *app.GitopiaProxy, storageManager *storage.Manager) PackfileUpdatedEventHandler {
 	return PackfileUpdatedEventHandler{
-		gc:           g,
-		pinataClient: pinataClient,
+		gc:             g,
+		storageManager: storageManager,
 	}
 }
 
@@ -110,17 +112,37 @@ func (h *PackfileUpdatedEventHandler) Handle(ctx context.Context, eventBuf []byt
 		return errors.WithMessage(err, "event parse error")
 	}
 
-	for _, event := range events {
-		if err := h.Process(ctx, event); err != nil {
-			// Log error and continue processing other events
-			logger.FromContext(ctx).WithFields(logrus.Fields{
-				"repository_id": event.RepositoryId,
-				"new_cid":       event.NewCid,
-				"old_cid":       event.OldCid,
-			}).WithError(err).Error("failed to process PackfileUpdatedEvent")
-		}
+	// Process events concurrently with a limited number of workers
+	// Limit to 5 concurrent packfile operations
+	maxWorkers := 5
+	if len(events) < maxWorkers {
+		maxWorkers = len(events)
 	}
 
+	semaphore := make(chan struct{}, maxWorkers)
+	var wg sync.WaitGroup
+
+	for _, event := range events {
+		wg.Add(1)
+		go func(e PackfileUpdatedEvent) {
+			defer wg.Done()
+			
+			// Acquire semaphore
+			semaphore <- struct{}{}
+			defer func() { <-semaphore }()
+
+			if err := h.Process(ctx, e); err != nil {
+				// Log error and continue processing other events
+				logger.FromContext(ctx).WithFields(logrus.Fields{
+					"repository_id": e.RepositoryId,
+					"new_cid":       e.NewCid,
+					"old_cid":       e.OldCid,
+				}).WithError(err).Error("failed to process PackfileUpdatedEvent")
+			}
+		}(event)
+	}
+
+	wg.Wait()
 	return nil
 }
 
@@ -132,29 +154,37 @@ func (h *PackfileUpdatedEventHandler) Process(ctx context.Context, event Packfil
 		"deleted":       event.Deleted,
 	}).Info("processing packfile updated event")
 
-	if event.NewCid != "" && !event.Deleted {
-		cacheDir := viper.GetString("GIT_REPOS_DIR")
-
-		// cache repo
-		utils.LockRepository(event.RepositoryId)
-		defer utils.UnlockRepository(event.RepositoryId)
-
-		if err := utils.CacheRepository(event.RepositoryId, cacheDir); err != nil {
+	if event.NewCid != "" && !event.Deleted && h.storageManager.HasProviders() {
+		refCount, err := h.gc.StorageCidReferenceCount(ctx, event.NewCid)
+		if err != nil {
+			logger.FromContext(ctx).WithError(err).Error("failed to get reference count")
 			return err
 		}
 
-		packfilePath := path.Join(cacheDir, fmt.Sprintf("%v.git/objects/pack/", event.RepositoryId), event.NewName)
-		resp, err := h.pinataClient.PinFile(ctx, packfilePath, filepath.Base(packfilePath))
-		if err != nil {
-			logger.FromContext(ctx).WithError(err).Error("failed to pin file to Pinata")
-			// Don't fail the process, just log the error
-		} else {
-			logger.FromContext(ctx).WithFields(logrus.Fields{
-				"repository_id": event.RepositoryId,
-				"packfile_name": event.NewName,
-				"cid":           event.NewCid,
-				"pinata_id":     resp.Data.ID,
-			}).Info("successfully pinned to Pinata")
+		if refCount == 1 {
+			cacheDir := viper.GetString("GIT_REPOS_DIR")
+
+			// cache repo
+			utils.RLockRepository(event.RepositoryId)
+			defer utils.RUnlockRepository(event.RepositoryId)
+
+			if err := utils.CacheRepository(event.RepositoryId, cacheDir); err != nil {
+				return err
+			}
+
+			packfilePath := path.Join(cacheDir, fmt.Sprintf("%v.git/objects/pack/", event.RepositoryId), event.NewName)
+			name := fmt.Sprintf("packfiles/%s", filepath.Base(packfilePath))
+			err := h.storageManager.PinFile(ctx, packfilePath, name)
+			if err != nil {
+				logger.FromContext(ctx).WithError(err).Error("failed to pin file to external storage")
+				// Don't fail the process, just log the error
+			} else {
+				logger.FromContext(ctx).WithFields(logrus.Fields{
+					"repository_id": event.RepositoryId,
+					"packfile_name": event.NewName,
+					"cid":           event.NewCid,
+				}).Info("successfully pinned to external storage")
+			}
 		}
 	}
 
@@ -166,16 +196,17 @@ func (h *PackfileUpdatedEventHandler) Process(ctx context.Context, event Packfil
 			return err
 		}
 
-		if refCount == 0 {
-			err := h.pinataClient.UnpinFile(ctx, event.OldName)
+		if refCount == 0 && h.storageManager.HasProviders() {
+			name := fmt.Sprintf("packfiles/%s", event.OldName)
+			err := h.storageManager.UnpinFile(ctx, name)
 			if err != nil {
-				logger.FromContext(ctx).WithError(err).Error("failed to unpin file from Pinata")
+				logger.FromContext(ctx).WithError(err).Error("failed to unpin file from external storage")
 			}
 			logger.FromContext(ctx).WithFields(logrus.Fields{
 				"repository_id": event.RepositoryId,
 				"old_name":      event.OldName,
 				"old_cid":       event.OldCid,
-			}).Info("unpinned file from Pinata")
+			}).Info("unpinned file from external storage")
 		}
 	}
 
@@ -187,20 +218,21 @@ func (h *PackfileUpdatedEventHandler) Process(ctx context.Context, event Packfil
 			return err
 		}
 
-		if refCount == 0 {
-			err := h.pinataClient.UnpinFile(ctx, event.NewName)
+		if refCount == 0 && h.storageManager.HasProviders() {
+			name := fmt.Sprintf("packfiles/%s", event.NewName)
+			err := h.storageManager.UnpinFile(ctx, name)
 			if err != nil {
 				logger.FromContext(ctx).WithFields(logrus.Fields{
 					"repository_id": event.RepositoryId,
 					"name":          event.NewName,
 					"cid":           event.NewCid,
-				}).WithError(err).Error("failed to unpin packfile from Pinata")
+				}).WithError(err).Error("failed to unpin packfile from external storage")
 			} else {
 				logger.FromContext(ctx).WithFields(logrus.Fields{
 					"repository_id": event.RepositoryId,
 					"name":          event.NewName,
 					"cid":           event.NewCid,
-				}).Info("successfully unpinned packfile from Pinata")
+				}).Info("successfully unpinned packfile from external storage")
 			}
 		}
 	}