+++ hooks/gitopia-pre-receive/config/config_testing.go
@@ -1,7 +1,0 @@
-//go:build testing
-
-package config
-
-const (
-	GRPCHost = "host.docker.internal:9100"
-)