+++ internal/app/handler/pr/diff.go
@@ -34,8 +34,8 @@ func PullDiffHandler(w http.ResponseWriter, r *http.Request) {
 		}
 
 		// cache repo
-		utils.LockRepository(body.BaseRepositoryID)
-		defer utils.UnlockRepository(body.BaseRepositoryID)
+		utils.RLockRepository(body.BaseRepositoryID)
+		defer utils.RUnlockRepository(body.BaseRepositoryID)
 
 		cacheDir := viper.GetString("GIT_REPOS_DIR")
 		if err := utils.CacheRepository(body.BaseRepositoryID, cacheDir); err != nil {
@@ -45,8 +45,8 @@ func PullDiffHandler(w http.ResponseWriter, r *http.Request) {
 
 		if body.HeadRepositoryID != body.BaseRepositoryID {
 			// cache repo
-			utils.LockRepository(body.HeadRepositoryID)
-			defer utils.UnlockRepository(body.HeadRepositoryID)
+			utils.RLockRepository(body.HeadRepositoryID)
+			defer utils.RUnlockRepository(body.HeadRepositoryID)
 
 			if err := utils.CacheRepository(body.HeadRepositoryID, cacheDir); err != nil {
 				http.Error(w, err.Error(), http.StatusInternalServerError)